# src/OnDeploy/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Err:
    $ref: '#/definitions/models.Res'
  models.GITDetail:
    properties:
      address:
        type: string
      git_pass:
        type: string
      git_path:
        type: string
      git_port:
        type: integer
      git_repo:
        type: string
      git_user:
        type: string
      name:
        type: string
      port:
        description: |-
          Username string `json:"username,omitempty"`
          Password string `json:"password,omitempty"`
        type: integer
    type: object
  models.NFSInfo:
    properties:
      address:
        type: string
      path:
        type: string
      port:
        type: integer
    type: object
  models.NFSPath:
    properties:
      acl:
        type: string
      address:
        type: string
      name:
        type: string
      param:
        items:
          type: string
        type: array
      path:
        type: string
      port:
        description: |-
          Username string `json:"username,omitempty"`
          Password string `json:"password,omitempty"`
        type: integer
    type: object
...
  models.Res:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  models.ServerDetail:
    properties:
      address:
        type: string
      name:
        type: string
      port:
        description: |-
          Username string `json:"username,omitempty"`
          Password string `json:"password,omitempty"`
        type: integer
    type: object
...
info:
  contact: {}
  description: 应用交付部署服务API
  license: {}
  title: 应用交付部署服务
  version: "1.0"
paths:
  /app/git/install:
    post:
      consumes:
      - application/json
      description: Git服务安装
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git服务安装
      tags:
      - Git服务
  /app/git/repo/add:
    put:
      consumes:
      - application/json
      description: Git仓库创建
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git仓库创建
      tags:
      - Git服务
  /app/git/repo/del:
    delete:
      consumes:
      - application/json
      description: Git仓库删除
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git仓库删除
      tags:
      - Git服务
  /app/git/user/add:
    post:
      consumes:
      - application/json
      description: Git用户添加
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git用户添加
      tags:
      - Git服务
  /app/git/user/del:
    delete:
      consumes:
      - application/json
      description: Git用户删除
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git用户删除
      tags:
      - Git服务
  /app/git/user/update:
    put:
      consumes:
      - application/json
      description: Git用户密码修改
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.GITDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: Git用户密码修改
      tags:
      - Git服务
  /app/nfs/install:
    post:
      consumes:
      - application/json
      description: NFS服务安装
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.ServerDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: NFS服务安装
      tags:
      - NFS服务
  /app/nfs/path/add:
    put:
      consumes:
      - application/json
      description: NFS服务创建
      parameters:
      - description: nfs
        in: body
        name: nfs
        required: true
        schema:
          $ref: '#/definitions/models.NFSPath'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: NFS服务创建
      tags:
      - NFS服务
  /app/nfs/path/del:
    delete:
      consumes:
      - application/json
      description: NFS服务删除
      parameters:
      - description: nfs
        in: body
        name: nfs
        required: true
        schema:
          $ref: '#/definitions/models.NFSInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: NFS服务删除
      tags:
      - NFS服务
  /app/nfs/path/lst/{address}:
    get:
      consumes:
      - application/json
      description: NFS服务查看
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: NFS服务查看
      tags:
      - NFS服务
...
schemes:
- http
- https
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"